+++ R/vellumplot-package.R
@@ -2,7 +2,7 @@
 "_PACKAGE"
 
 #' @importFrom rlang enquos eval_tidy quo_is_null quo_get_expr `%||%`
-#' @importFrom vellum vl_scene push pop draw viewport grid_layout unit gpar
+#' @importFrom vellum vl_scene push pop draw vl_viewport grid_layout vl_unit vl_gpar
 #' @importFrom vellum points_grob lines_grob segments_grob rect_grob text_grob
 #' @importFrom vellum grobwidth grobheight render as_vellum_scene
 NULL